refactor : 컨트롤러와 엔트리 포인트 변경
