+++ src/main/java/lotto/Application.java
@@ -1,0 +1,9 @@
+package lotto;
+
+import lotto.controller.LottoController;
+
+public class Application {
+    public static void main(String[] args) {
+        LottoController.run();
+    }
+}
