+++ src/main/java/lotto/controller/LottoController.java
@@ -1,4 +1,4 @@
-package lotto;
+package lotto.controller;
 
 import lotto.domain.Accountant;
 import lotto.domain.LottoNumber;
@@ -8,9 +8,8 @@ import lotto.domain.Money;
 import lotto.view.InputView;
 import lotto.view.OutputView;
 
-public class LottoApplication {
-    public static void main(String[] args) {
-        // TODO : 10줄 이내로 변경
+public class LottoController {
+    public static void run() {
         OutputView.inputMoneyInstruction();
         Money money = new Money(InputView.getInput());
         OutputView.ticketAmountInstruction(money);
